Move lifi api types to separate file
